Split files into errors, tokenizer, parser, codegen
